refactor(chat): extract chat-interface into maintainable components and hooks
Refactored the monolithic chat-interface component (1,629 lines) into a
well-organized architecture for improved maintainability and testability.

**New Hooks:**
- use-inline-errors: Error state management with deduplication
- use-attachment-upload: File upload logic with state tracking
- use-stream-lifecycle: Stream event tracking and performance metrics

**New View Components:**
- chat-prompt-input: Reusable prompt input eliminating code duplication
- chat-new-session-view: Centered empty state for new chat sessions
- chat-existing-session-view: Thread view with messages and input

**New Utilities:**
- message-validation: Attachment validation against model capabilities
- attachment-processing: Data transformation and type conversion utilities

**Type Fixes:**
- Fixed ChatStatus imports from "ai" package
- Corrected feedback mutation signatures
- Updated ValidationError to use literal `retryable: false` type
- Migrated from PromptInputAttachmentItem to PromptInputAttachmentPayload
- Fixed PromptInputProps to exclude both onSubmit and onError from HTMLAttributes

**Results:**
- Reduced chat-interface.tsx by ~39% (1,629 → 1,000 lines)
- Separated concerns: state management, UI, business logic
- Maintained full type safety without any usage
- All builds, type checks, and lints pass
